requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.18.0
fonttools==4.39.4
fparser==0.1.1
fqdn==1.5.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.14.0
isoduration==20.11.0
jax==0.4.11
jaxlib==0.4.11
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpickle==3.0.1
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kaleido==0.2.1
kiwisolver==1.4.4
langchain==0.0.219
langchainplus-sdk==0.0.17
llvmlite==0.40.1
-e git+https://github.com/anthony-zhou/climate_code_conversion.git@c478f655e241a0ce9f494da694fe8622c96a437c#egg=local_translation
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.1
notebook==7.0.0
notebook_shim==0.2.3
numba==0.57.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.15.0
pluggy==1.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
promptlayer==0.1.90
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pygraphviz==1.11
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pyvis==0.3.2
PyYAML==6.0
pyzmq==24.0.1
referencing==0.30.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.9.2
scipy==1.10.1
Send2Trash==1.8.2
setuptools-scm==7.1.0
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
yarl==1.9.2
zipp==3.15.0
